Return flags for fixed variables (#7)
The `fx` tag indicates variables whose values are intended to be held
fixed. This is needed if we are to run the problem as intended. An
alternative would be to adjust `lo` and `hi`, but this approach makes it
easier for "consumers" to use unconstrained optimization with a subset
of variables.